Merge back

+++ packages/cli/src/api/e2e.api.ts
@@ -5,50 +5,65 @@
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
 /* eslint-disable @typescript-eslint/naming-convention */
 import { Router } from 'express';
+import type { Request } from 'express';
 import bodyParser from 'body-parser';
 import { v4 as uuid } from 'uuid';
+import { Container } from 'typedi';
 import config from '@/config';
 import * as Db from '@/Db';
 import type { Role } from '@db/entities/Role';
+import { RoleRepository } from '@db/repositories';
 import { hashPassword } from '@/UserManagement/UserManagementHelper';
 import { eventBus } from '@/eventbus/MessageEventBus/MessageEventBus';
+import { License } from '../License';
+import { LICENSE_FEATURES } from '@/constants';
 
 if (process.env.E2E_TESTS !== 'true') {
 	console.error('E2E endpoints only allowed during E2E tests');
 	process.exit(1);
 }
 
-const tablesToTruncate = [
-	'auth_identity',
-	'auth_provider_sync_history',
-	'event_destinations',
-	'shared_workflow',
-	'shared_credentials',
-	'webhook_entity',
-	'workflows_tags',
-	'credentials_entity',
-	'tag_entity',
-	'workflow_statistics',
-	'workflow_entity',
-	'execution_entity',
-	'settings',
-	'installed_packages',
-	'installed_nodes',
-	'user',
-	'role',
-];
+const enabledFeatures = {
+	[LICENSE_FEATURES.SHARING]: true, //default to true here instead of setting it in config/index.ts for e2e
+	[LICENSE_FEATURES.LDAP]: false,
+	[LICENSE_FEATURES.SAML]: false,
+	[LICENSE_FEATURES.LOG_STREAMING]: false,
+	[LICENSE_FEATURES.ADVANCED_EXECUTION_FILTERS]: false,
+	[LICENSE_FEATURES.VERSION_CONTROL]: false,
+};
+
+type Feature = keyof typeof enabledFeatures;
+
+Container.get(License).isFeatureEnabled = (feature: Feature) => enabledFeatures[feature] ?? false;
+
+const tablesNotToTruncate = ['sqlite_sequence'];
 
 const truncateAll = async () => {
-	const { connection } = Db;
-	for (const table of tablesToTruncate) {
-		await connection.query(
-			`DELETE FROM ${table}; DELETE FROM sqlite_sequence WHERE name=${table};`,
-		);
+	const connection = Db.getConnection();
+	const allTables: Array<{ name: string }> = await connection.query(
+		"SELECT name FROM sqlite_master WHERE type='table';",
+	);
+
+	// Disable foreign key constraint checks
+	await connection.query('PRAGMA foreign_keys = OFF;');
+
+	for (const { name: table } of allTables) {
+		try {
+			if (tablesNotToTruncate.includes(table)) continue;
+			await connection.query(
+				`DELETE FROM ${table}; DELETE FROM sqlite_sequence WHERE name=${table};`,
+			);
+		} catch (error) {
+			console.warn('Dropping Table for E2E Reset error: ', error);
+		}
 	}
+
+	// Re-enable foreign key constraint checks
+	await connection.query('PRAGMA foreign_keys = ON;');
 };
 
 const setupUserManagement = async () => {
-	const { connection } = Db;
+	const connection = Db.getConnection();
 	await connection.query('INSERT INTO role (name, scope) VALUES ("owner", "global");');
 	const instanceOwnerRole = (await connection.query(
 		'SELECT last_insert_rowid() as insertId',
@@ -78,7 +93,7 @@ const setupUserManagement = async () => {
 };
 
 const resetLogStreaming = async () => {
-	config.set('enterprise.features.logStreaming', false);
+	enabledFeatures[LICENSE_FEATURES.LOG_STREAMING] = false;
 	for (const id in eventBus.destinations) {
 		await eventBus.removeDestination(id);
 	}
@@ -100,13 +115,7 @@ e2eController.post('/db/setup-owner', bodyParser.json(), async (req, res) => {
 		return;
 	}
 
-	const globalRole = await Db.collections.Role.findOneOrFail({
-		select: ['id'],
-		where: {
-			name: 'owner',
-			scope: 'global',
-		},
-	});
+	const globalRole = await Container.get(RoleRepository).findGlobalOwnerRoleOrFail();
 
 	const owner = await Db.collections.User.findOneByOrFail({ globalRoleId: globalRole.id });
 
@@ -127,7 +136,14 @@ e2eController.post('/db/setup-owner', bodyParser.json(), async (req, res) => {
 	res.writeHead(204).end();
 });
 
-e2eController.post('/enable-feature/:feature', async (req, res) => {
-	config.set(`enterprise.features.${req.params.feature}`, true);
-	res.writeHead(204).end();
-});
+e2eController.patch(
+	'/feature/:feature',
+	bodyParser.json(),
+	async (req: Request<{ feature: Feature }>, res) => {
+		const { feature } = req.params;
+		const { enabled } = req.body;
+
+		enabledFeatures[feature] = enabled === undefined || enabled === true;
+		res.writeHead(204).end();
+	},
+);
